feat: merge webhook-client into rest
