+++ package.json
@@ -18,9 +18,7 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.26.0",
-    "@sapphire/eslint-config": "4.3.8",
     "@types/node": "18.11.18",
-    "@types/rimraf": "3.0.2",
     "dotenv": "16.0.3",
     "eslint": "8.31.0",
     "eslint-config-neon": "^0.1.40",
