+++ packages/guilded.js/lib/index.ts
@@ -6,17 +6,5 @@ export * from "./cache";
 export * from "./constants";
 export * from "./util";
 export * from "./typings";
-export {
-  Embed as WebhookEmbed,
-  EmbedStructure as WebhookEmbedStructure,
-  WebhookClient,
-  WebhookExecuteResponse,
-  enforcedMessageStructure,
-  COLORS,
-  resolveColor,
-  parseMessage,
-  parseToMessage,
-  parsedMessage,
-  parsedTextResponse,
-} from "@guildedjs/webhook-client";
+export * from "@guildedjs/rest";
 export * from "@discordjs/collection";
