+++ packages/guilded.js/package.json
@@ -22,7 +22,6 @@
   "dependencies": {
     "@discordjs/collection": "^1.3.0",
     "@guildedjs/rest": "workspace:*",
-    "@guildedjs/webhook-client": "workspace:*",
     "@guildedjs/ws": "workspace:*",
     "typed-emitter": "2.1.0"
   },
