--- packages/webhook-client/README.md
@@ -1,24 +1,0 @@
-# `@guildedjs/webhook-client`
-
-[![GitHub](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/guildedjs/guilded.js/blob/main/LICENSE)
-[![npm](https://img.shields.io/npm/v/@guildedjs/webhook-client?color=crimson&logo=npm)](https://www.npmjs.com/package/@guildedjs/webhook-client)
-[![CI workflows](https://github.com/guildedjs/guilded.js/actions/workflows/ci.yml/badge.svg)](https://github.com/guildedjs/guilded.js/actions/workflows/ci.yml)
-
-Library agnostic client for interfacing with Guilded webhooks.
-
-> ⚠️ If you intend to use `guilded.js` you won't need to install this, as it's already a dependency of that package.
-
-## Installation
-
-You can install this package from [NPM](https://www.npmjs.com/package/@guildedjs/webhook-client)
-
--   `npm install @guildedjs/webhook-client`
--   `yarn add @guildedjs/webhook-client`
-
-## Contributing
-
-Please see the main [README.md](https://github.com/guildedjs/guilded.js) for info on how to contribute to this package or the other `@guildedjs` packages.
-
-## LICENSE
-
-Licensed under the [MIT License](https://github.com/guildedjs/guilded.js/blob/main/LICENSE)
