--- packages/webhook-client/__tests__/file-sending/src/index.ts
@@ -1,8 +1,0 @@
-import { readFileSync } from "node:fs";
-import { WebhookClient } from "../../../lib/index";
-
-const client = new WebhookClient("");
-
-void (async (): Promise<unknown> => {
-    return client.send({ files: [{ name: "bbb.png", content: readFileSync("./bbb.png") }] });
-})();
