--- packages/webhook-client/lib/consts.ts
@@ -1,36 +1,0 @@
-/**
- * Copyright 2015 - 2021 Amish Shah
- * Copyrights licensed under the Apache License 2.0, https://github.com/discordjs/discord.js/blob/master/LICENSE
- * Taken from https://github.com/discordjs/discord.js/blob/stable/src/util/Constants.js
- */
-export const COLORS: { [key: string]: number } = {
-    AQUA: 0x1abc9c,
-    BLUE: 0x3498db,
-    DARKER_GREY: 0x7f8c8d,
-    DARK_AQUA: 0x11806a,
-    DARK_BLUE: 0x206694,
-    DARK_BUT_NOT_BLACK: 0x2c2f33,
-    DARK_GOLD: 0xc27c0e,
-    DARK_GREEN: 0x1f8b4c,
-    DARK_GREY: 0x979c9f,
-    DARK_NAVY: 0x2c3e50,
-    DARK_ORANGE: 0xa84300,
-    DARK_PURPLE: 0x71368a,
-    DARK_RED: 0x992d22,
-    DARK_VIVID_PINK: 0xad1457,
-    DEFAULT: 0x000000,
-    GOLD: 0xf1c40f,
-    GREEN: 0x2ecc71,
-    GREY: 0x95a5a6,
-    GREYPLE: 0x99aab5,
-    GUILDED: 0xf5c400,
-    LIGHT_GREY: 0xbcc0c0,
-    LUMINOUS_VIVID_PINK: 0xe91e63,
-    NAVY: 0x34495e,
-    NOT_QUITE_BLACK: 0x23272a,
-    ORANGE: 0xe67e22,
-    PURPLE: 0x9b59b6,
-    RED: 0xe74c3c,
-    WHITE: 0xffffff,
-    YELLOW: 0xffff00,
-} as const;
