--- packages/webhook-client/lib/index.ts
@@ -1,5 +1,0 @@
-export * from "./consts";
-export * from "./Embed";
-export * from "./messageUtil";
-export * from "./util";
-export * from "./WebhookClient";
