--- packages/webhook-client/package.json
@@ -1,73 +1,0 @@
-{
-  "name": "@guildedjs/webhook-client",
-  "version": "0.4.5",
-  "description": "Structures and clients for webhooks in Guilded.",
-  "author": "Zaid \"Nico\" <contact@nico.engineer>",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "types": "types/index.d.ts",
-  "typedoc": {
-    "entryPoint": "lib/index.ts"
-  },
-  "scripts": {
-    "build": "tsc && gen-esm-wrapper . ./dist/index.mjs",
-    "build:typecheck": "tsc --noEmit",
-    "prepublishOnly": "rimraf dist/ && rimraf types/ && pnpm run build"
-  },
-  "devDependencies": {
-    "typescript": "4.9.4"
-  },
-  "dependencies": {
-    "@guildedjs/guilded-api-typings": "workspace:*",
-    "@guildedjs/rest": "workspace:*",
-    "form-data": "^4.0.0"
-  },
-  "contributors": [
-    {
-      "name": "Zaid \"Nico\"",
-      "email": "contact@nico.engineer",
-      "url": "https://github.com/zaida04"
-    },
-    {
-      "name": "Skillz4Killz",
-      "email": "guildedjs@drskillz.33mail.com",
-      "url": "https://github.com/Skillz4Killz"
-    },
-    {
-      "name": "Uhuh \"Dylan\"",
-      "email": "dylan@panku.io",
-      "url": "https://github.com/uhuh"
-    },
-    {
-      "name": "DaStormer",
-      "email": "dastormer@stormdevelopmentz.xyz",
-      "url": "https://github.com/DaStormer"
-    }
-  ],
-  "exports": {
-    ".": {
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
-    },
-    "./": "./"
-  },
-  "keywords": [
-    "guilded",
-    "guildedjs",
-    "guilded.js",
-    "guilded-api"
-  ],
-  "files": [
-    "dist",
-    "types"
-  ],
-  "homepage": "https://github.com/guildedjs/guilded.js/tree/main/packages/rest#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/guildedjs/guilded.js.git"
-  },
-  "bugs": {
-    "url": "https://github.com/guildedjs/guilded.js/issues"
-  },
-  "gitHead": "eee38a19e0bfa812d7136cc78a6bc99e0b402b0c"
-}
